prevent filtering of localized results from twitter

+++ src/org/loklak/data/DAO.java
@@ -578,7 +578,7 @@ public class DAO {
     }
     
     public static class SearchLocalMessages {
-        public long hits;
+        public int hits;
         public Timeline timeline;
         public Map<String, List<Map.Entry<String, Long>>> aggregations;
 
@@ -617,7 +617,7 @@ public class DAO {
                 }
                 // get response
                 SearchResponse response = request.execute().actionGet();
-                this.hits = response.getHits().getTotalHits();
+                this.hits = (int) response.getHits().getTotalHits();
                         
                 // evaluate search result
                 //long totalHitCount = response.getHits().getTotalHits();
